Extract ARM specific parts of timer code.
This diff shoudl be a no-op, but enables adding timer support for other platforms
in future.
