+++ examples/rp2040/minimal/src/main.rs
@@ -66,7 +66,7 @@ fn main() -> ! {
 
     // Configure the systick timer for 1kHz ticks at the default ROSC speed of
     // _roughly_ 6 MHz.
-    lilos::time::initialize_sys_tick(&mut cp.SYST, 6_000_000);
+    lilos::cortex_m_timer::initialize_sys_tick(&mut cp.SYST, 6_000_000);
     // Set up and run the scheduler with a single task.
     lilos::exec::run_tasks(
         &mut [blink],  // <-- array of tasks
