+++ examples/stm32f4/uart-echo/src/main.rs
@@ -92,7 +92,7 @@ fn main() -> ! {
     device::GPIOD.moder().modify(|w| w.set_moder(15, Moder::OUTPUT));
 
     // Set up and run the scheduler.
-    lilos::time::initialize_sys_tick(&mut cp.SYST, CLOCK_HZ);
+    lilos::cortex_m_timer::initialize_sys_tick(&mut cp.SYST, CLOCK_HZ);
     lilos::exec::run_tasks_with_idle(
         &mut [heartbeat, echo],
         lilos::exec::ALL_TASKS,
