+++ examples/stm32h7/minimal/src/main.rs
@@ -61,7 +61,7 @@ fn main() -> ! {
 
     // Configure the systick timer for 1kHz ticks at 64MHz (the speed the CPU is
     // running at when it leaves reset).
-    lilos::time::initialize_sys_tick(&mut cp.SYST, 64_000_000);
+    lilos::cortex_m_timer::initialize_sys_tick(&mut cp.SYST, 64_000_000);
     // Set up and run the scheduler with a single task.
     lilos::exec::run_tasks(
         &mut [blink],  // <-- array of tasks
